Fix heroku entry point error
