+++ Procfile
@@ -1,2 +1,1 @@
 web gunicorn run:app
-setuptools==1.4
