+++ run.py
@@ -1,9 +1,10 @@
 import os
 
-from app.v1.app import create_app
+from app import create_app
+
+env_name = os.getenv('FLASK_ENV')
+app = create_app(env_name)
 
 if __name__ == '__main__':
-    env_name = os.getenv('FLASK_ENV')
-    app = create_app(env_name)
     # app.config['SESSION_TYPE'] = 'filesystem'
     app.run()
